# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File Storage System (User, Folder, File)",
    "description": "Сводная OpenAPI документация по трём сервисам: User Service, Folder Service, File Service.",
    "version": "1.0.0"
  },
  "tags": [
    { "name": "auth", "description": "Авторизация (JWT)" },
    { "name": "users", "description": "Управление пользователями" },
    { "name": "folders", "description": "Управление папками" },
    { "name": "files", "description": "Управление файлами" }
  ],
  "paths": {
    "/auth/token": {
      "post": {
        "tags": [ "auth" ],
        "summary": "Получить JWT-токен по логину/паролю",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно получен токен",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "400": {
            "description": "Неверный логин или пароль"
          }
        }
      }
    },


    "/users": {
      "post": {
        "tags": [ "users" ],
        "summary": "Создать нового пользователя",
        "security": [ { "bearerAuth": [ ] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь создан",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRead" }
              }
            }
          },
          "400": {
            "description": "Логин уже занят или другая ошибка"
          }
        }
      },
      "get": {
        "tags": [ "users" ],
        "summary": "Получить список всех пользователей",
        "security": [ { "bearerAuth": [ ] } ],
        "responses": {
          "200": {
            "description": "Список пользователей",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserRead" }
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [ "users" ],
        "summary": "Получить пользователя по ID",
        "security": [ { "bearerAuth": [ ] } ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRead" }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        }
      }
    },


    "/folders": {
      "post": {
        "tags": [ "folders" ],
        "summary": "Создать новую папку",
        "security": [ { "bearerAuth": [ ] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FolderCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Папка создана",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FolderRead" }
              }
            }
          }
        }
      },
      "get": {
        "tags": [ "folders" ],
        "summary": "Получить список всех папок",
        "security": [ { "bearerAuth": [ ] } ],
        "responses": {
          "200": {
            "description": "Список папок",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FolderRead" }
                }
              }
            }
          }
        }
      }
    },
    "/folders/{folder_id}": {
      "get": {
        "tags": [ "folders" ],
        "summary": "Получить папку по её ID",
        "security": [ { "bearerAuth": [ ] } ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о папке",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FolderRead" }
              }
            }
          },
          "404": {
            "description": "Папка не найдена"
          }
        }
      },
      "patch": {
        "tags": [ "folders" ],
        "summary": "Обновить (переименовать) папку",
        "security": [ { "bearerAuth": [ ] } ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FolderUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Папка обновлена",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FolderRead" }
              }
            }
          },
          "404": {
            "description": "Папка не найдена"
          }
        }
      },
      "delete": {
        "tags": [ "folders" ],
        "summary": "Удалить папку",
        "security": [ { "bearerAuth": [ ] } ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Папка успешно удалена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Папка не найдена"
          }
        }
      }
    },


    "/files": {
      "post": {
        "tags": [ "files" ],
        "summary": "Загрузить (создать) новый файл",
        "security": [ { "bearerAuth": [ ] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FileCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Файл успешно загружен",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileRead" }
              }
            }
          }
        }
      },
      "get": {
        "tags": [ "files" ],
        "summary": "Получить список всех файлов (метаданные)",
        "security": [ { "bearerAuth": [ ] } ],
        "responses": {
          "200": {
            "description": "Список файлов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FileRead" }
                }
              }
            }
          }
        }
      }
    },
    "/files/{file_id}": {
      "get": {
        "tags": [ "files" ],
        "summary": "Получить метаданные конкретного файла",
        "security": [ { "bearerAuth": [ ] } ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Метаданные файла",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileRead" }
              }
            }
          },
          "404": {
            "description": "Файл не найден"
          }
        }
      },
      "delete": {
        "tags": [ "files" ],
        "summary": "Удалить файл по ID",
        "security": [ { "bearerAuth": [ ] } ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Файл удалён",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Файл не найден"
          }
        }
      }
    },
    "/files/{file_id}/content": {
      "get": {
        "tags": [ "files" ],
        "summary": "Получить содержимое файла",
        "security": [ { "bearerAuth": [ ] } ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Содержимое файла (base64 или текст)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileContent" }
              }
            }
          },
          "404": {
            "description": "Файл не найден"
          }
        }
      }
    }
  },


  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "TokenRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": [ "username", "password" ]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "default": "bearer" }
        },
        "required": [ "access_token" ]
      },

      "UserCreate": {
        "type": "object",
        "properties": {
          "login": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": [ "login", "password" ]
      },
      "UserRead": {
        "type": "object",
        "properties": {
          "user_id": { "type": "integer" },
          "login": { "type": "string" }
        }
      },

      "FolderCreate": {
        "type": "object",
        "properties": {
          "user_id": { "type": "integer" },
          "name": { "type": "string" }
        },
        "required": [ "user_id", "name" ]
      },
      "FolderRead": {
        "type": "object",
        "properties": {
          "folder_id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "FolderUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" }
        },
        "required": [ "name" ]
      },

      "FileCreate": {
        "type": "object",
        "properties": {
          "folder_id": { "type": "integer" },
          "filename": { "type": "string" },
          "content": { "type": "string" }
        },
        "required": [ "folder_id", "filename" ]
      },
      "FileRead": {
        "type": "object",
        "properties": {
          "file_id": { "type": "integer" },
          "folder_id": { "type": "integer" },
          "filename": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "FileContent": {
        "type": "object",
        "properties": {
          "content": { "type": "string" }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": [ ]
    }
  ]
}